S20 - Project 11 - HR Resume Screening Assistance/Resume Screening  Assistance Project - Source Code/requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
unstructured
pinecone-client
pypdf
sentence_transformers